# [14.0] sale_stock_available_to_promise_release: improve '_get_next_replenishment_date' perf