Checking ongoing states instead of finished ones (done and canceled) helps PostgreSQL to perform the filter on far less lines (done lines are increasing a lot over time), speeding up the query to get the next replenishment date.

Some figures based on a product having ~82k moves over 5M in the table. The query has been improved by a factor of 50.

**Before**
```sql
 Limit  (cost=1000.49..36535.12 rows=1 width=12) (actual time=1233.129..1236.150 rows=1 loops=1)
   ->  Nested Loop  (cost=1000.49..320812.16 rows=9 width=12) (actual time=1233.128..1236.148 rows=1 loops=1)
         Join Filter: (stock_move.picking_type_id = stock_picking_type.id)
         Rows Removed by Join Filter: 7
         ->  Gather Merge  (cost=1000.49..320786.59 rows=162 width=16) (actual time=1232.167..1236.087 rows=2 loops=1)
               Workers Planned: 4
               Workers Launched: 2
               ->  Parallel Index Scan using stock_move_date_index on stock_move  (cost=0.43..319767.24 rows=40 width=16) (actual time=860.442..1226.830 rows=3 loops=3)
                     Filter: ((((state)::text <> ALL ('{cancel,done}'::text[])) OR (state IS NULL)) AND (product_id = 7449) AND (company_id = 1))
                     Rows Removed by Filter: 1731876
         ->  Materialize  (cost=0.00..6.25 rows=8 width=4) (actual time=0.011..0.027 rows=4 loops=2)
               ->  Seq Scan on stock_picking_type  (cost=0.00..6.21 rows=8 width=4) (actual time=0.018..0.048 rows=6 loops=1)
                     Filter: ((company_id = 1) AND ((code)::text = 'incoming'::text))
                     Rows Removed by Filter: 104
 Planning Time: 0.496 ms
 Execution Time: 1236.199 ms
```

**After**
```sql
 Limit  (cost=6776.42..6776.42 rows=1 width=12) (actual time=25.782..25.785 rows=1 loops=1)
   ->  Sort  (cost=6776.42..6776.44 rows=9 width=12) (actual time=25.780..25.782 rows=1 loops=1)
         Sort Key: stock_move.date
         Sort Method: top-N heapsort  Memory: 25kB
         ->  Nested Loop  (cost=6569.58..6776.37 rows=9 width=12) (actual time=25.627..25.770 rows=6 loops=1)
               Join Filter: (stock_move.picking_type_id = stock_picking_type.id)
               Rows Removed by Join Filter: 681
               ->  Bitmap Heap Scan on stock_move  (cost=6569.58..6750.80 rows=162 width=16) (actual time=25.390..25.522 rows=116 loops=1)
                     Recheck Cond: (((product_id = 7449) AND (company_id = 1) AND ((state)::text = ANY ('{partially_available,confirmed,assigned,waiting,draft}'::text[]))) OR (state IS NULL))
                     Filter: ((product_id = 7449) AND (company_id = 1))
                     Heap Blocks: exact=104
                     ->  BitmapOr  (cost=6569.58..6569.58 rows=162 width=0) (actual time=25.367..25.368 rows=0 loops=1)
                           ->  Bitmap Index Scan on stock_move_product_location_index  (cost=0.00..6567.96 rows=162 width=0) (actual time=25.356..25.356 rows=125 loops=1)
                                 Index Cond: ((product_id = 7449) AND (company_id = 1) AND ((state)::text = ANY ('{partially_available,confirmed,assigned,waiting,draft}'::text[])))
                           ->  Bitmap Index Scan on stock_move_state_index  (cost=0.00..1.54 rows=1 width=0) (actual time=0.008..0.008 rows=0 loops=1)
                                 Index Cond: (state IS NULL)
               ->  Materialize  (cost=0.00..6.25 rows=8 width=4) (actual time=0.000..0.001 rows=6 loops=116)
                     ->  Seq Scan on stock_picking_type  (cost=0.00..6.21 rows=8 width=4) (actual time=0.028..0.057 rows=6 loops=1)
                           Filter: ((company_id = 1) AND ((code)::text = 'incoming'::text))
                           Rows Removed by Filter: 104
 Planning Time: 0.647 ms
 Execution Time: 25.862 ms
```